Create AbstractLine
